+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Add dtrx to help with extracting archives via the CLI.
 - Add support for building with a MBR partition table.
 - Change SDDM user profile picture to a better quality image.
+- Change Pacman to use 'curl-static' for package downloads.
 
 ## 4.1.0 - The Dual-Boot Update